+++ src/main.cpp
@@ -9,13 +9,13 @@ class $modify(BIRGarageLayer, GJGarageLayer) {
     bool init() override {
         if (!GJGarageLayer::init()) return false;
 
-        if (ButtonColorSettingValue::randomColor == ButtonColor::Random) {
+        if (ButtonColorSettingValue::BUTTON_COLOR == ButtonColor::Random) {
             auto color = Mod::get()->getSettingValue<ButtonColor>("randomize-button-color");
-            ButtonColorSettingValue::randomColor = color == ButtonColor::Random ? static_cast<ButtonColor>(BetterIconRandomizer::randomNumber(1, 5)) : color;
+            ButtonColorSettingValue::BUTTON_COLOR = color == ButtonColor::Random ? static_cast<ButtonColor>(BetterIconRandomizer::randomNumber(1, 7)) : color;
         }
 
         auto randomizeBtn = CCMenuItemExt::createSpriteExtra(CCSprite::createWithSpriteFrameName(fmt::format("BIR_randomBtn_{:02d}_001.png"_spr,
-            static_cast<int>(ButtonColorSettingValue::randomColor)).c_str()), [this](auto) { BIRSelectPopup::create(this)->show(); });
+            static_cast<int>(ButtonColorSettingValue::BUTTON_COLOR)).c_str()), [this](auto) { BIRSelectPopup::create(this)->show(); });
         randomizeBtn->setID("select-randomize-button"_spr);
         auto shardsMenu = getChildByID("shards-menu");
         shardsMenu->addChild(randomizeBtn);